iOS Compatibility: Build fixes for Comparator utilities
- Remove unused Comparator import from CharmEffect.java
- Fix Zone.java Float.compare for color order weight
- Add verifyTransitivity stub to ComparatorUtil for error reporting

Fixes compilation errors and completes iOS build.

+++ forge-game/src/main/java/forge/game/zone/Zone.java
@@ -61,8 +61,8 @@ public class Zone implements java.io.Serializable, Iterable<Card> {
             int cmc = Integer.compare(c1.getCMC(), c2.getCMC());
             if (cmc != 0) return cmc;
 
-            // Then compare by color order weight
-            int color = Integer.compare(c1.getColor().getOrderWeight(), c2.getColor().getOrderWeight());
+            // Then compare by color order weight (returns float)
+            int color = Float.compare(c1.getColor().getOrderWeight(), c2.getColor().getOrderWeight());
             if (color != 0) return color;
 
             // Then compare by name
